+++ modules/vpn-server-configuration/src/terraform.tf
@@ -1,10 +1,14 @@
 terraform {
-  required_version = "~> 1.5"
+  required_version = "~> 1.11"
 
   required_providers {
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "~> 3.1"
+    }
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.85"
+      version = "~> 4.23"
     }
   }
 }