+++ modules/kubernetes-cluster/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 4.0
+* Updated the required Terraform version from `1.5` to `1.11`.
+* Updated the AzureRM Terraform provider version from `3.x` to `4.x`.
+* Changed the default CNI plugin from `kubenet` to `azure`.
+* Renamed `subnet_id` variable to `vnet_subnet_id`.
+
 ## 3.0
 * Added `os_disk_size_gb`, `os_disk_type` and `os_sku` variables to enable greater default node pool customisation.
 * Changed default node pool VM OS disk type to `Ephemeral`.