+++ modules/storage-account/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## 1.2
+## 2.0
+* Updated the required Terraform version from `1.5` to `1.11`.
+* Updated the AzureRM Terraform provider version from `3.x` to `4.x`.
 
+## 1.2
 * Added support for static website.
 
 ## 1.1
-
 * Added `primary_access_key` and `secondary_access_key` outputs.