+++ modules/virtual-network/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 3.0
+* Updated the required Terraform version from `1.5` to `1.11`.
+* Updated the AzureRM Terraform provider version from `3.x` to `4.x`.
+
 ## 2.0
 * Removed `dns_servers` variable to enable the use of the [azurerm_virtual_network_dns_servers](https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/virtual_network_dns_servers) resource instead.